src/FilesToData.ApiPy/requirements.txt
azure-functions
azure-storage-queue
pypdf
requests
